myone ready for lounch

+++ app/src/main/java/com/humaclab/selliscope_myone/model/AddNewOrder.java
@@ -17,7 +17,7 @@ public class AddNewOrder implements Serializable {
         @SerializedName("outlet_id")
         public String outletId;
         @SerializedName("discount")
-        public int discount;
+        public Double discount;
         @SerializedName("order_date")
         public String date;
         @SerializedName("products")
